filled out civil war + header
I finished filling out the information for the civil war page. I will probably eventually go back and add images for the important figures.

I also did work on the header. I'm worried about how if we want to update the links on the header we have to edit every single html file. I relocated all of the css for the header into "header.css" but I don't know if there's anything we can do about the html.